Let the user trigger a full-screen refresh
My opinion is that a full-screen refresh is appropriate when switching
between "significantly different" sections of an app... for example,
Folly would do it when starting or finishing a story.

This code tracks the current "route" - a term for a section of the app,
by metaphor with web routing - and does a full-screen flash every time
the route changes. I don't hate it so far!